небольшой рефакторинг
