+++ src/main/java/ilya/project/telegrambot/bot/commands/Command.java
@@ -1,0 +1,10 @@
+package ilya.project.telegrambot.bot.commands;
+
+import org.telegram.telegrambots.meta.api.objects.Update;
+import org.telegram.telegrambots.meta.bots.AbsSender;
+
+public interface Command {
+    String commandName();
+
+    void apply(Update update, AbsSender sender);
+}
