+++ src/main/java/ilya/project/telegrambot/bot/commands/CommandChain.java
@@ -1,9 +1,10 @@
-package ilya.project.telegrambot.commands;
+package ilya.project.telegrambot.bot.commands;
 
 import org.telegram.telegrambots.meta.api.objects.Update;
 import org.telegram.telegrambots.meta.bots.AbsSender;
 
-public interface Command {
-    String commandName();
+public interface CommandChain {
+    void next(CommandChain next);
+
     void apply(Update update, AbsSender sender);
 }
