+++ src/main/java/ilya/project/telegrambot/bot/config/BotRegistration.java
@@ -1,6 +1,6 @@
-package ilya.project.telegrambot.config;
+package ilya.project.telegrambot.bot.config;
 
-import ilya.project.telegrambot.TelegramBot;
+import ilya.project.telegrambot.bot.TelegramBot;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.event.ContextRefreshedEvent;
@@ -19,9 +19,12 @@ public class BotRegistration {
     @EventListener({ContextRefreshedEvent.class})
     public void init() {
         try {
+            log.debug("Re-init Telegram bot");
             TelegramBotsApi telegramBotsApi = new TelegramBotsApi(DefaultBotSession.class);
             telegramBotsApi.registerBot(bot);
+            log.debug("Bot was successfully re-registered");
         } catch (TelegramApiException e) {
+            log.error("error with re-register Telegram bot!");
             log.error(e.getMessage());
         }
     }
