--- src/main/java/ilya/project/telegrambot/config/BotConfig.java
@@ -1,14 +1,0 @@
-package ilya.project.telegrambot.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.telegram.telegrambots.bots.DefaultBotOptions;
-
-@Configuration
-public class BotConfig {
-
-    @Bean
-    public DefaultBotOptions botOptions() {
-        return new DefaultBotOptions();
-    }
-}
